+++ addon/components/loading-alert.ts
@@ -1,0 +1,3 @@
+//@ts-expect-error ember-appuniversum uses vanilla javascript
+import AuAlertComponent from '@appuniversum/ember-appuniversum/components/au-alert';
+export default class LoadingAlertComponent extends AuAlertComponent {}