+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removal of the `entries` attribute from the table-of-contents prosemirror node. The node can now generate it's own outline in its `serialize` method.
 ### Dependencies
 - Bumps `@codemirror/view` from 6.12.0 to 6.16.0
+- Bumps `@lblod/ember-rdfa-editor` from 4.2.0 to 5.1.0
 
 ## [10.0.0] - 2023-08-08
 ### Changed