+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - GN-4442: template comments can move up and down over the *whole* document
 
+### Added
+- GN4470: ability to specifly single/multi-select per codelist instance
+
 ## [11.0.0] - 2023-08-22
 
 ### Added